v6.1.9

+++ bus-core/src/main/java/org/aoju/bus/core/lang/Murmur.java
@@ -2,7 +2,7 @@
  *                                                                               *
  * The MIT License (MIT)                                                         *
  *                                                                               *
- * Copyright (c) 2015-2020 aoju.org and other contributors.                      *
+ * Copyright (c) 2015-2021 aoju.org and other contributors.                      *
  *                                                                               *
  * Permission is hereby granted, free of charge, to any person obtaining a copy  *
  * of this software and associated documentation files (the "Software"), to deal *
@@ -33,7 +33,7 @@ import java.io.Serializable;
  * Murmur3 32bit、64bit、128bit 哈希算法实现
  *
  * @author Kimi Liu
- * @version 6.1.8
+ * @version 6.1.9
  * @since JDK 1.8+
  */
 public class Murmur implements Serializable {
@@ -335,7 +335,7 @@ public class Murmur implements Serializable {
             case 2:
                 k1 ^= (long) (data[tailStart + 1] & 0xff) << 8;
             case 1:
-                k1 ^= (long) (data[tailStart] & 0xff);
+                k1 ^= data[tailStart] & 0xff;
                 k1 *= C1;
                 k1 = Long.rotateLeft(k1, R1);
                 k1 *= C2;
